requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
billiard==4.2.0
black==23.12.0
boto3==1.34.19
botocore==1.34.19
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.105.0
filetype==1.2.0
flake8==6.1.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
jmespath==1.0.1
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
PyMuPDF==1.23.19
PyMuPDFb==1.23.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.13
